requirements.txt
alembic==1.7.7
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
blis==0.7.6
Boruta==0.3
brotlipy==0.7.0
cachetools==5.0.0
catalogue==1.0.0
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625814692085/work
chardet @ file:///tmp/build/80754af9/chardet_1607706775000/work
click==8.0.4
cloudpickle==2.0.0
colorlover==0.3.0
conda==4.10.3
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262147379/work
cryptography @ file:///tmp/build/80754af9/cryptography_1616767007030/work
cssutils==2.4.0
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.6
databricks-cli==0.16.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docker==5.0.3
efficientnet-pytorch==0.7.1
entrypoints==0.4
executing==0.8.3
fastai==2.5.3
fastcore==1.3.29
fastdownload==0.0.5
fastprogress==1.0.2
Flask==2.0.3
fonttools==4.30.0
funcy==1.17
future==0.18.2
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
gunicorn==20.1.0
htmlmin==0.1.12
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
ImageHash==4.2.1
imbalanced-learn==0.7.0
importlib-metadata==4.11.3
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.0
kmodes==0.11.1
lightgbm==3.3.2
llvmlite==0.38.0
Mako==1.2.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
mlflow==1.24.0
mlxtend==0.19.0
multimethod==1.7
murmurhash==1.0.6
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.7.1
nltk==3.7
notebook==6.4.10
numba==0.55.1
numexpr==2.8.1
numpy==1.19.5
packaging==21.3
pandas==1.4.1
pandas-profiling==3.1.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
phik==0.12.1
pickleshare==0.7.5
Pillow==9.0.1
plac==1.1.3
plotly==5.6.0
preshed==3.0.6
prometheus-client==0.13.1
prometheus-flask-exporter==0.19.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pycaret==2.3.6
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.9.0
pydeck==0.7.1
Pygments==2.11.2
pyLDAvis==3.2.2
Pympler==1.0.1
pynndescent==0.5.6
pyod==0.9.8
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==22.3.0
querystring-parser==1.2.4
regex==2022.3.15
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016711199/work
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.1
spacy==2.3.7
SQLAlchemy==1.4.32
sqlparse==0.4.2
srsly==1.0.5
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.7.0
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
terminado==0.13.3
testpath==0.6.0
textblob==0.17.1
thinc==7.4.5
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.11.2
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets==5.1.1
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
umap-learn==0.5.2
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
validators==0.18.2
visions==0.7.4
wasabi==0.9.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wordcloud==1.8.1
yellowbrick==1.3.post1
zipp==3.7.0
